deprecate GameScreen and rename to GameScreenOld
