+++ core/src/com/obstacleavoid/game/ObstacleAvoidGame.kt
@@ -3,10 +3,10 @@ package com.obstacleavoid.game
 import com.badlogic.gdx.Application
 import com.badlogic.gdx.Game
 import com.badlogic.gdx.Gdx
-import com.badlogic.gdx.graphics.GL20
 import com.badlogic.gdx.graphics.Texture
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
-import com.obstacleavoid.game.screen.GameScreen
+import com.obstacleavoid.game.screen.game.GameScreen
+import com.obstacleavoid.game.screen.game.GameScreenOld
 
 class ObstacleAvoidGame : Game() {
     var batch: SpriteBatch? = null
