+++ core/src/com/obstacleavoid/game/screen/game/GameScreen.kt
@@ -1,0 +1,33 @@
+package com.obstacleavoid.game.screen.game
+
+import com.badlogic.gdx.Screen
+
+class GameScreen : Screen {
+    override fun hide() {
+        TODO("Not yet implemented")
+    }
+
+    override fun show() {
+        TODO("Not yet implemented")
+    }
+
+    override fun render(delta: Float) {
+        TODO("Not yet implemented")
+    }
+
+    override fun pause() {
+        TODO("Not yet implemented")
+    }
+
+    override fun resume() {
+        TODO("Not yet implemented")
+    }
+
+    override fun resize(width: Int, height: Int) {
+        TODO("Not yet implemented")
+    }
+
+    override fun dispose() {
+        TODO("Not yet implemented")
+    }
+}
